meantetris: add falling square.
Mean Tetris is simply a tetris game where the shapes do not fall at
random. Instead choosing the shapes to be the worst possible fit for the
current board.
Explanation of the classes implemented:
* TetrisGame represent the game (with score and current falling shape)
* GameState represents the game board (only fallen shapes)
* Shape are the falling shapes, and the shape choosing is left to the
  ShapeFactory.
* TetrisController represent the possible contollers used to move the
  pieces (could be kayboard, mouse or possibly handling network for a
  remote based controls)
* TetrisDisplay represents the possible displays (for now in the
  Terminal, but could be in a separate window, ...)
For now, only Terminal display is implemented, we only have the Square
shape and the controller does nothing (the shape is falling).